+++ source/SlideOutMenu.js
@@ -14,8 +14,10 @@ const PureComponent = React.PureComponent || React.Component
 
 export default class ContextAwareSlideoutMenu extends PureComponent {
 	slideOutMenu = createRef()
+
 	show = () => this.slideOutMenu.current.show()
 	hide = () => this.slideOutMenu.current.hide()
+
 	render() {
 		return (
 			<PageAndMenuContext.Consumer>
@@ -79,6 +81,10 @@ class SlideoutMenu extends PureComponent
 		setTogglerCooldown: PropTypes.func.isRequired,
 		getTogglerNode: PropTypes.func.isRequired,
 
+		// If `scrollIntoView` is `true` (which is the default)
+		// then these two are gonna define the delay after which it scrolls into view.
+		expandAnimationDuration : PropTypes.number.isRequired,
+
 		// CSS style object
 		style : PropTypes.object,
 
@@ -90,13 +96,15 @@ class SlideoutMenu extends PureComponent
 	{
 		// isOpen : false,
 		anchor : 'left',
-		fullscreen : false
+		fullscreen : false,
+
+		expandAnimationDuration : 220
 	}
 
 	state = {
-		// Should not be `undefined` because it's compared to
-		// `show` argument in `.toggle(show)`.
-		show: false
+		// Initializing it to `false` is required for `if (expand === expanded)`.
+		//
+		expanded: false
 	}
 
 	container = createRef()
@@ -105,13 +113,10 @@ class SlideoutMenu extends PureComponent
 	componentDidMount()
 	{
 		const { registerMenu, menuRef } = this.props
-		const { show } = this.state
 
 		this.unregister = registerMenu
 		({
-			hide    : () => this.setState({ show: false }),
 			toggle  : this.toggle,
-			isShown : () => this.state.show,
 			element : () => this.container.current,
 			menu    : () => menuRef ? menuRef.current : this.container.current
 		})
@@ -123,32 +128,170 @@ class SlideoutMenu extends PureComponent
 	componentWillUnmount() {
 		this.unregister()
 		// window.removeEventListener('popstate', this.hide)
+
+		if (this.cancelProcess) {
+			this.cancelProcess()
+		}
 	}
 
-	toggle = (show, onAfterToggle) => {
+	toggle = (expand, { onCollapsed, onExpanded } = {}) => {
+		// This code was copy-pasted from `source/Expandable.js`.
+
 		const { onCollapse, onExpand } = this.props
-		if (show === this.state.show) {
-			return
+
+		const { expanded } = this.state
+
+		// If no `expand` argument provided then just toggle.
+		const isExpandedState = (expanded && this.process !== 'unexpand') ||
+			(this.process === 'expand')
+		if (expand === undefined) {
+			expand = !isExpandedState
 		}
-		if (show === undefined) {
-			show = !this.state.show
+
+		// Don't collapse if already collapsed.
+		// Don't expand if already expanded.
+
+		if (expand === isExpandedState) {
+			return Promise.resolve()
 		}
-		if (show) {
-			this.onFocusOutRef.current.listenToTouches()
-			onExpand && onExpand()
+
+		// if (this.isToggling) {
+		// 	return Promise.resolve()
+		// }
+
+		if (this.cancelProcess) {
+			this.cancelProcess()
+		}
+
+		let cancelled
+
+		const reset = () => {
+			this.process = undefined
+			this.processPhase = undefined
+			this.cancelProcess = undefined
+		}
+
+		const nextPhase = (phase) => {
+			if (!cancelled) {
+				this.processPhase = phase
+				return true
+			}
+		}
+
+		function endProcess() {
+			nextPhase()
+			reset()
+		}
+
+		const phaseCancellers = {}
+		function onCancelPhase(phase, handler) {
+			phaseCancellers[phase] = handler
+		}
+
+		this.cancelProcess = () => {
+			if (phaseCancellers[this.processPhase]) {
+				phaseCancellers[this.processPhase]()
+			}
+			cancelled = true
+			reset()
+		}
+
+		if (expand) {
+			// Expand.
+			this.process = 'expand'
+			onCancelPhase('scheduleExpand', () => clearTimeout(this.expandTimeout))
+			return new Promise((resolve) => {
+				if (!nextPhase('render')) {
+					return
+				}
+				this.setState({
+					shouldRender: true
+				},
+				// Without an artificial delay for some reason the CSS "expand" animation won't play.
+				// Perhaps a browser decides to optimize two subsequent renders
+				// and doesn't render "pre-expanded" and "expanded" states separately.
+				// Even with a 0ms delay it would randomly play/not-play the expand animation.
+				() => {
+					if (!nextPhase('scheduleExpand')) {
+						return
+					}
+					if (onExpand) {
+						onExpand()
+					}
+					// Using `requestAnimationFrame()` instead of `setTimeout()`
+					// because otherwise there would be a weird and strange delay.
+					this.expandTimeout = setTimeout(() => {
+						if (!nextPhase('expand')) {
+							return
+						}
+						this.expandTimeout = undefined
+						this.setState({ expanded: true }, () => {
+							if (!nextPhase('expanded')) {
+								return
+							}
+							if (onExpanded) {
+								onExpanded()
+							}
+							if (this.onFocusOutRef.current) {
+								this.onFocusOutRef.current.listenToTouches()
+							}
+							endProcess()
+							resolve()
+						})
+					}, 0)
+				})
+			})
 		} else {
-			this.onFocusOutRef.current.stopListeningToTouches()
-			onCollapse && onCollapse()
+			// Un-Expand.
+			this.process = 'unexpand'
+			onCancelPhase('unexpanded', () => clearTimeout(this.waitUnForExpandAnimationTimer))
+			nextPhase('unexpand')
+			if (this.onFocusOutRef.current) {
+				this.onFocusOutRef.current.stopListeningToTouches()
+			}
+			if (onCollapse) {
+				onCollapse()
+			}
+			// Set `expanded` to `false` to play the collapse CSS animation.
+			// Once that animation is finished remove
+			// the contents of the `<Expanded/>` from DOM.
+			return new Promise((resolve) => {
+				this.setState({ expanded: false }, () => {
+					if (!nextPhase('unexpanded')) {
+						return
+					}
+					if (onCollapsed) {
+						onCollapsed()
+					}
+					const waitForUnExpandAnimation = () => {
+						const { expandAnimationDuration } = this.props
+						return new Promise((resolve) => {
+							this.waitUnForExpandAnimationTimer = setTimeout(resolve, expandAnimationDuration * 1.1)
+						})
+					}
+					waitForUnExpandAnimation().then(() => {
+						if (!nextPhase('unrender')) {
+							return
+						}
+						this.setState({ shouldRender: false }, endProcess)
+					})
+					// `resolve()` doesn't wait for `removeFromDOMAfterCollapsed`
+					// because other components use it like `.toggle().then(focus)`
+					// where it shouldn't wait for the final phases.
+					resolve()
+				})
+			})
 		}
-		this.setState({ show }, onAfterToggle)
+	}
+
+	show = () => {
+		const { toggleMenu } = this.props
+		toggleMenu(true)
 	}
 
 	hide = () => {
 		const { toggleMenu } = this.props
-		const { show } = this.state
-		if (show) {
-			toggleMenu()
-		}
+		toggleMenu(false)
 	}
 
 	onKeyDown = (event) => {
@@ -160,7 +303,7 @@ class SlideoutMenu extends PureComponent
 			// Collapse on "Escape".
 			case 27:
 				event.preventDefault()
-				return toggleMenu()
+				return toggleMenu(false)
 		}
 	}
 
@@ -192,10 +335,18 @@ class SlideoutMenu extends PureComponent
 			toggleMenu,
 			registerMenu,
 			setTogglerCooldown,
+			expandAnimationDuration,
 			...rest
 		} = this.props
 
-		const { show } = this.state
+		const {
+			shouldRender,
+			expanded
+		} = this.state
+
+		if (!shouldRender) {
+			return null
+		}
 
 		// ARIA menu notes:
 		// https://www.w3.org/TR/wai-aria-practices/examples/menu-button/menu-button-links.html
@@ -214,7 +365,7 @@ class SlideoutMenu extends PureComponent
 				<div
 					{ ...rest }
 					ref={ this.container }
-					aria-hidden={ !show }
+					aria-hidden={ !expanded }
 					tabIndex={ -1 }
 					onKeyDown={ this.onKeyDown }
 					className={ classNames(
@@ -229,7 +380,7 @@ class SlideoutMenu extends PureComponent
 							'rrui__slideout-menu--top'        : anchor === 'top',
 							'rrui__slideout-menu--bottom'     : anchor === 'bottom',
 							'rrui__slideout-menu--fullscreen' : fullscreen,
-							'rrui__slideout-menu--expanded'   : show
+							'rrui__slideout-menu--expanded'   : expanded
 						}
 					) }>
 					{ children }